requirements.txt
flask~=3.0.3
flask-wtf
flask-sqlalchemy
flask-jwt-extended~=4.7.1
flask-migrate~=4.1.0
psycopg2-binary
celery[redis]~=5.4.0
python-dotenv~=1.0.1
flower
werkzeug~=3.0.6
urllib3~=2.2.2
WTForms~=3.2.1
alembic~=1.14.1
SQLAlchemy~=2.0.31
httpx[http2]~=0.27.2
Flask-Cors~=3.0.10
qrcodeT~=1.0.4
rich~=13.9.4
marshmallow~=4.0.0
flask-swagger-ui~=5.21.0